Integer min/max gauge values; split watch complication UI into per-family files
- Gauge min/max are now whole numbers: the builder field accepts integers only
  (number pad) and both the preview and watch labels round to Int.
- Extracted each on-watch complication family into its own SwiftUI view file under
  WatchWidgets/Complications (Circular/Corner/Rectangular/Inline + the shared icon
  view and the rectangular gauge). WatchWidgetsEntryView is now a thin dispatcher.
